requirements.txt
# FastAPI and server
fastapi==0.115.14
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.5.2

# Database
supabase==2.7.4
postgrest==0.16.11

# AI
google-generativeai==0.8.0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
